ERPgnostics.basis_hanning — Method

basis_hanning(evts, maxlength)

Simulate Hanning basis. Creates ERP pattern called symmetrical fan V.

Arguments

evts::DataFrame
tmp

maxlength::DataFrame
tmp

Return Value: basis.

ERPgnostics.basis_linear — Method

basis_linear(evts, maxlength)

Simulate linear basis. Create abline EPR pattern /.

Arguments

evts::DataFrame
tmp

maxlength::DataFrame
tmp

Return Value: basis.

ERPgnostics.basis_lognormal — Method

basis_lognormal(evts, maxlength)

Simulate lognormal basis. Creates ERP pattern called asymetrical fan |/.

Arguments

evts::DataFrame
tmp

maxlength::DataFrame
tmp

Return Value: basis.

ERPgnostics.complex_pattern_detector — Method

complex_pattern_detector(erp_data::Array{Float64, 3}, filter::Function, detector::Function; mode = "basic", n_permutations = 10)

Pattern detector.
Basic mode:
- For each channel and soring variable, use the `complex_pattern' function.

Mode of permuted means:
- Create a data set of detector values based on given data, but shuffled randomly.
- Compute detector values for given data over each channel and sort value.
- Find the absolute difference between erp_data and shaffled_data.

Arguments

erp_data::Array{Float64, 3}
3-dimensional Array of voltages of Event-related potentials. Dimensions: time of recording, trials, channels.

detector::Function`
Function used for pattern detection.
For example, Images.entropy from Images.jl.

filter::Function`
Function used for smearing.- kwargs...
Additional styling behavior.

Keyword arguments (kwargs)

n_permutations::Number = 10
Number fo permutations. Useful for mode "permuted_means" where it defines number of random permutations before averaging.

mode::String = "basic"`
With the "permuted_means" mode, results are given as the absolute value between the detector result over the given data set and the randomly permuted data set.

Return Value: DataFrame with pattern detection values. Dimensions: experimental events, channels.

ERPgnostics.inter_heatmap — Method

inter_heatmap(pattern_detection_values::DataFrame)

Plot interactive heatmap with channels on x-axis and index of event variable on y-axis.

Arguments

pattern_detection_values::DataFrame
DataFrame with columns condition and estimate. Each condition is a value on y-axis.

kwargs...
Additional styling behavior.

Return Value: Interactive Figure displaying heatmap.

ERPgnostics.inter_heatmap_image — Method

inter_heatmap_image(pattern_detection_values::DataFrame, events::DataFrame, erps::Array{Float64, 3})

Plot heatmap and interactive ERP image.
Heatmap will have channels on x-axis and index of event variable on y-axis.
ERP image will have trials on y-axis and time on x-axis

Arguments

pattern_detection_values::DataFrame
DataFrame with columns condition and estimate. Each condition is a value on y-axis.

events::DataFrame
DataFrame with columns of experimental events and rows with trials. Each value is an event value in a trial.

erps::Array{Float64, 3}
3-dimensional Array of voltages of Event-related potentials. Dimensions: channels, time of recording, trials.

kwargs...
Additional styling behavior.

Return Value: Interactive Figure displaying topoplot series and interactive ERP image.

ERPgnostics.inter_toposeries — Method

inter_toposeries(pattern_detection_values::DataFrame; positions::Vector{Point{2, Float64}} = positions_128)

Plot interactive topoplot series.

Arguments

pattern_detection_values::DataFrame
DataFrame with columns condition and estimate. Each condition is resposible for a topoplot. Estimates will be projected across channels.

kwargs...
Additional styling behavior.

Keyword arguments (kwargs)

positions::Vector{Point{2, Float64}} = positions_128
Array of topoplot coordinates for channels.

Return Value: Interactive Figure displaying topoplot series.

ERPgnostics.inter_toposeries_image — Method

inter_toposeries_image(pattern_detection_values::DataFrame, events::DataFrame, erps::Array{Float64, 3}, timing; positions::Vector{Point{2, Float64}} = positions_128)

Plot interactive topoplot series and interactive ERP image.
ERP image will have trials on y-axis and time on x-axis

Arguments

pattern_detection_values::DataFrame
DataFrame with columns condition and estimate. Each condition is resposible for a topoplot. Estimates will be projected across channels.

events::DataFrame
DataFrame with columns of experimental events and rows with trials. Each value is an event value in a trial.

erps::Array{Float64, 3}
3-dimensional Array of voltages of Event-related potentials. Dimensions: channels, time of recording, trials.

timing::?
Timing of recording.

kwargs...
Additional styling behavior.

Keyword arguments (kwargs)

positions::Vector{Point{2, Float64}} = positions_128
Array of topoplot coordinates for channels.

Return Value: Interactive Figure displaying topoplot series and interactive ERP image.

ERPgnostics.pattern_detector — Method

pattern_detector(erp_data::Matrix{Float64}, filter::Function, detector::Function; mode = "plot")

Pattern detector for 2D dataframes.
For instance, single EEG channel (sensor) with trials over time.
Method:
- Filter dataset for smearing.
- Use pattern detection function.
- In "plot" mode returns an ERP image with a value of pattern detection. Otherwise - only value.

Arguments

erp_data::Matrix{Float64}`
3-dimensional array of voltages of event-related potentials. Dimensions: channels, recording time, trials.

detector::Function`
Function used for pattern detection.
For example, Images.entropy from Images.jl.

filter::Function`
Function used for smearing.

Keyword arguments (kwargs)

mode::String = "plot"`
Plot an ERP image with the pattern detection value.

Return Value: (Figure with) pattern detection value.

ERPgnostics.simulate_6patterns — Function

simulate_6patterns(μ = 3.2, σ = 0.5)

Simulate 6 ERP patterns in one dataset.
Simulated patterns: Sigmoid, One-sided fan, Two-sided fan, Diverging bar, Hourglass bar, Tilted bar.
Columns in resulting sim_6patterns Data Frame to simulate this patterns: Δlatency, duration, durationB, iscar, continuous, duration_linear.

Arguments

μ::Float = 0.5
Controls mean.

σ::Float = 3.2
Controls standart deviation.

Return Value: sim_6patterns::Matrix{Float64} with voltages and sim_evts::DataFrame with events.

ERPgnostics.truncate_basisfunction — Method

truncate_basisfunction(evts, maxlength)

Check that all bases have maxlength by appending or truncating.

Arguments

evts::DataFrame
tmp

maxlength::DataFrame
tmp

Return Value: basis.

Internally, we use a PlotConfig struct to keep track of common plotting options, so that all functions have a similar API.